PortalBackend/requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
decorator==4.0.11
Django==3.0.5
django-cors-headers==3.5.0
djangorestframework==3.12.1
djongo==1.3.3
dnspython==2.0.0
future==0.18.2
idna==2.10
imageio==2.1.2
InstagramAPI==1.0.2
moviepy==0.2.3.2
numpy==1.19.4
oauthlib==3.1.0
Pillow==8.0.1
pymongo==3.11.0
PySocks==1.7.1
python-twitter==3.5
pytz==2020.1
requests==2.11.1
requests-oauthlib==1.3.0
requests-toolbelt==0.7.0
selenium==3.141.0
soupsieve==2.0.1
sqlparse==0.2.4
stem==1.8.0
torrequest==0.1.0
tqdm==4.11.2
urllib3==1.26.2
